requirements.txt
Django==4.2.7
django_environ==0.11.2
djangorestframework==3.14.0
Requests==2.31.0
pillow==10.2.0